// frontend/src/gen/TaskManagement.ts
export interface TaskResponse {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: string title = 2
     */
    title: string;
    /**
     * @generated from protobuf field: string description = 3
     */
    description: string;
    /**
     * @generated from protobuf field: int32 project_id = 4
     */
    projectId: number;
    /**
     * @generated from protobuf field: int32 assignee_id = 5
     */
    assigneeId: number;
    /**
     * @generated from protobuf field: taskmanagement.TaskStatus status = 6
     */
    status: TaskStatus;
    /**
     * @generated from protobuf field: taskmanagement.TaskPriority priority = 7
     */
    priority: TaskPriority;
    /**
     * @generated from protobuf field: int32 created_by_id = 8
     */
    createdById: number;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp created_at = 9
     */
    createdAt?: Timestamp;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp updated_at = 10
     */
    updatedAt?: Timestamp;
}
// ...
export enum TaskStatus {
    /**
     * @generated from protobuf enum value: TASK_STATUS_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: TASK_STATUS_TODO = 1;
     */
    TODO = 1,
    /**
     * @generated from protobuf enum value: TASK_STATUS_IN_PROGRESS = 2;
     */
    IN_PROGRESS = 2,
    /**
     * @generated from protobuf enum value: TASK_STATUS_REVIEW = 3;
     */
    REVIEW = 3,
    /**
     * @generated from protobuf enum value: TASK_STATUS_DONE = 4;
     */
    DONE = 4
}
/**
 * @generated from protobuf enum taskmanagement.TaskPriority
 */
export enum TaskPriority {
    /**
     * @generated from protobuf enum value: TASK_PRIORITY_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: TASK_PRIORITY_LOW = 1;
     */
    LOW = 1,
    /**
     * @generated from protobuf enum value: TASK_PRIORITY_MEDIUM = 2;
     */
    MEDIUM = 2,
    /**
     * @generated from protobuf enum value: TASK_PRIORITY_HIGH = 3;
     */
    HIGH = 3,
    /**
     * @generated from protobuf enum value: TASK_PRIORITY_CRITICAL = 4;
     */
    CRITICAL = 4
}
// ...
export const TaskResponse = new TaskResponse$Type();